requirements.txt
# AI-Gauge dependencies
openai>=1.0.0
python-dotenv>=1.0.0
langgraph>=0.2.0
